fix(pgserve-v2#6): split daemon.js to stay under 1000-line limit
Resolves Group 6 FIX-FIRST gaps surfaced by reviewer-g6:
- src/daemon.js 1278 -> 597 lines
- New: src/daemon-control.js (390 lines, control-socket handlers)
- New: src/daemon-tcp.js (296 lines, --listen TCP path + token auth)
- New: src/daemon-shared.js (18 lines, shared utilities)
- WISH.md: clarify pgserve.toml allowlist deferred to v2.1
