+++ src/daemon.js
@@ -1,5 +1,5 @@
 /**
- * pgserve daemon — singleton control-socket server.
+ * pgserve daemon — singleton control-socket server (orchestrator).
  *
  * One process per host. Listens on a well-known Unix socket
  * (`$XDG_RUNTIME_DIR/pgserve/control.sock`, fallback `/tmp/pgserve/control.sock`),
@@ -10,12 +10,14 @@
  * the control socket. A second daemon invocation refuses with the live PID;
  * a stale lock (process gone) is cleaned up automatically on next boot.
  *
- * Wave-2 scope (this file): transport layer + fingerprint accept hook.
- *   - Group 3: every accept derives a kernel-rooted peer fingerprint and
- *     audits a `connection_routed` event.
- *   - Group 4 will rewrite the startup-message database parameter.
- *   - Group 5 will install GC sweep triggers on the daemon.
- *   - Group 6 will add the optional `--listen` TCP bind.
+ * Module layout (split for AGENTS.md §8 1000-line discipline):
+ *   - daemon.js (this file)   — class shell, lifecycle, lock, signal handlers,
+ *     listener wiring, public exports.
+ *   - daemon-control.js       — Unix accept hooks: handleSocketOpen/Data/Close/
+ *     Error, processStartupMessage, resolveTenantDatabase (Group 2 + Group 4).
+ *   - daemon-tcp.js           — Optional TCP accept hooks + token verify
+ *     (Group 6).
+ *   - daemon-shared.js        — flushPending helper shared by both paths.
  *
  * PR #24 invariants preserved:
  *   - `PostgresManager.start()` re-entry guard untouched.
@@ -29,31 +31,18 @@ import fs from 'fs';
 import path from 'path';
 import { EventEmitter } from 'events';
 import { PostgresManager } from './postgres.js';
-import { extractDatabaseName, extractApplicationName, rewriteDatabaseName, buildErrorResponse } from './protocol.js';
 import { createLogger } from './logger.js';
-import { handleControlAccept, initFingerprintFfi } from './fingerprint.js';
-import { configureAudit, audit, AUDIT_EVENTS } from './audit.js';
-import {
-  ensureMetaSchema,
-  verifyToken,
-  findRowByFingerprint,
-  recordDbCreated,
-  touchLastConnection,
-} from './control-db.js';
+import { initFingerprintFfi } from './fingerprint.js';
+import { configureAudit } from './audit.js';
+import { ensureMetaSchema } from './control-db.js';
 import { createAdminClient, writeAdminDiscovery, removeAdminDiscovery } from './admin-client.js';
-import { parseTcpAuth, hashToken } from './tokens.js';
 import {
-  resolveTenantDatabaseName,
   isFingerprintEnforcementDisabled,
   KILL_SWITCH_ENV,
 } from './tenancy.js';
-
-const PROTOCOL_VERSION_3 = 196608;
-const SSL_REQUEST_CODE = 80877103;
-const GSSAPI_REQUEST_CODE = 80877104;
-const CANCEL_REQUEST_CODE = 80877102;
-
-const MAX_STARTUP_BUFFER_SIZE = 1024 * 1024; // 1 MiB — same bound as router.js
+import { flushPending } from './daemon-shared.js';
+import { attachControlHandlers } from './daemon-control.js';
+import { attachTcpHandlers } from './daemon-tcp.js';
 
 /**
  * Resolve the directory that holds the daemon's control socket and pid lock.
@@ -107,11 +96,6 @@ export function isProcessAlive(pid) {
  *
  * Cleanup contract on failed acquisition is the caller's responsibility:
  * we never unlink the socket of a *live* peer.
- *
- * @param {object} opts
- * @param {string} opts.pidLockPath
- * @param {string} opts.socketPath
- * @param {object} [opts.logger]
  */
 export function acquirePidLock({ pidLockPath, socketPath, libpqCompatPath, logger }) {
   ensureDir(path.dirname(pidLockPath));
@@ -229,6 +213,9 @@ function sleepBlocking(ms) {
 
 /**
  * The daemon. Owns one PostgresManager and one Bun.listen({unix}) server.
+ * Accept-path methods (handleSocketOpen, handleTcpOpen, …) live in the
+ * daemon-control.js / daemon-tcp.js modules and are mixed into the
+ * prototype below.
  */
 export class PgserveDaemon extends EventEmitter {
   constructor(options = {}) {
@@ -544,365 +531,6 @@ export class PgserveDaemon extends EventEmitter {
     process.on('SIGHUP', onSignal);
   }
 
-  /**
-   * Route a client connection through to the underlying PG Unix socket.
-   *
-   * Per-accept fingerprint derivation is live: SO_PEERCRED → /proc walk →
-   * package.json hash, with a `connection_routed` audit emit. Fingerprint
-   * info is parked on socketState so Group 4 (database-per-fingerprint)
-   * can resolve the tenant DB without re-deriving on every byte.
-   */
-  handleSocketOpen(socket) {
-    let fingerprint = null;
-    try {
-      const opts = this._fingerprintAcceptOpts ? (this._fingerprintAcceptOpts(socket) || {}) : {};
-      fingerprint = handleControlAccept(socket, opts);
-    } catch (err) {
-      this.logger.warn?.(
-        { err: err?.message || String(err) },
-        'Failed to derive peer fingerprint on accept',
-      );
-    }
-    this.socketState.set(socket, {
-      buffer: null,
-      pgSocket: null,
-      dbName: null,
-      handshakeComplete: false,
-      startupInProgress: false,
-      pendingToPg: null,
-      pendingToClient: null,
-      fingerprint,
-    });
-    this.connections.add(socket);
-    if (fingerprint) {
-      this.emit('accept', { fingerprint, socket });
-    }
-  }
-
-  handleSocketData(socket, data) {
-    const state = this.socketState.get(socket);
-    if (!state) return;
-
-    if (state.handshakeComplete && state.pgSocket) {
-      if (state.pendingToPg) {
-        state.pendingToPg = Buffer.concat([state.pendingToPg, Buffer.from(data)]);
-        socket.pause();
-        return;
-      }
-      const written = state.pgSocket.write(data);
-      if (written < data.byteLength) {
-        state.pendingToPg = written === 0 ? Buffer.from(data) : Buffer.from(data.subarray(written));
-        socket.pause();
-      }
-      return;
-    }
-
-    const incomingSize = state.buffer ? state.buffer.length + data.byteLength : data.byteLength;
-    if (incomingSize > MAX_STARTUP_BUFFER_SIZE) {
-      this.logger.warn?.(
-        { incomingSize, limit: MAX_STARTUP_BUFFER_SIZE },
-        'Pre-handshake buffer exceeded limit — closing connection',
-      );
-      socket.end();
-      return;
-    }
-    if (state.buffer) {
-      state.buffer = Buffer.concat([state.buffer, Buffer.from(data)]);
-    } else {
-      state.buffer = Buffer.from(data);
-    }
-    this.processStartupMessage(socket, state).catch((err) => {
-      this.logger.error?.({ err: err.message }, 'processStartupMessage failed');
-      try { socket.end(); } catch { /* swallow */ }
-    });
-  }
-
-  async processStartupMessage(socket, state) {
-    if (state.startupInProgress) return;
-    const buffer = state.buffer;
-    if (!buffer || buffer.length < 8) return;
-
-    const messageLength = buffer.readUInt32BE(0);
-    if (buffer.length < messageLength) return;
-
-    const code = buffer.readUInt32BE(4);
-
-    if (code === SSL_REQUEST_CODE || code === GSSAPI_REQUEST_CODE) {
-      socket.write(Buffer.from('N'));
-      state.buffer = buffer.length > messageLength ? buffer.subarray(messageLength) : null;
-      return;
-    }
-
-    if (code === CANCEL_REQUEST_CODE) {
-      socket.end();
-      return;
-    }
-
-    if (code !== PROTOCOL_VERSION_3) {
-      this.logger.warn?.({ code }, 'Unsupported protocol version on control socket');
-      socket.end();
-      return;
-    }
-
-    const startupMessage = buffer.subarray(0, messageLength);
-    const requestedDb = extractDatabaseName(startupMessage);
-    state.dbName = requestedDb;
-    state.startupInProgress = true;
-
-    try {
-      // Group 4: resolve the tenant database for this peer's fingerprint.
-      // Auto-provisions on first connect, denies cross-tenant attempts
-      // (unless the kill switch is set), and rewrites the startup-message
-      // database parameter so the underlying PG sees the canonical name.
-      const resolution = await this.resolveTenantDatabase(state, requestedDb);
-      if (resolution.deny) {
-        const errFrame = buildErrorResponse({
-          severity: 'FATAL',
-          sqlstate: '28P01',
-          message: resolution.message,
-        });
-        try {
-          // Bun's TCPSocket.end(data) writes then closes atomically — using
-          // write()+end() can race the FIN past the data on some kernels and
-          // leave the peer waiting for AuthOK indefinitely.
-          socket.end(errFrame);
-        } catch { /* swallow */ }
-        this.emit('connection-denied', {
-          fingerprint: state.fingerprint?.fingerprint || null,
-          requested: requestedDb,
-          owned: resolution.ownedDatabaseName,
-        });
-        return;
-      }
-
-      const dbName = resolution.databaseName;
-      state.dbName = dbName;
-      const outgoingStartup = (dbName !== requestedDb)
-        ? rewriteDatabaseName(startupMessage, dbName)
-        : startupMessage;
-
-      if (this.autoProvision) {
-        await this.pgManager.createDatabase(dbName);
-      }
-
-      const pgSocketPath = this.pgManager.getSocketPath();
-      const daemon = this;
-      const pgHandler = {
-        data(_pgSocket, pgData) {
-          if (state.pendingToClient) {
-            state.pendingToClient = Buffer.concat([state.pendingToClient, Buffer.from(pgData)]);
-            _pgSocket.pause();
-            return;
-          }
-          const written = socket.write(pgData);
-          if (written < pgData.byteLength) {
-            state.pendingToClient = written === 0
-              ? Buffer.from(pgData)
-              : Buffer.from(pgData.subarray(written));
-            _pgSocket.pause();
-          }
-        },
-        open(pgSocket) {
-          pgSocket.write(outgoingStartup);
-          state.handshakeComplete = true;
-        },
-        close() {
-          try { socket.end(); } catch { /* swallow */ }
-        },
-        error(_pgSocket, error) {
-          daemon.logger.error?.({ dbName, err: error?.message || String(error) }, 'PG-side proxy socket error');
-          try { socket.end(); } catch { /* swallow */ }
-        },
-        drain(_pgSocket) {
-          if (state.pendingToPg) {
-            state.pendingToPg = flushPending(_pgSocket, state.pendingToPg);
-          }
-          if (!state.pendingToPg) {
-            socket.resume();
-          }
-        },
-      };
-
-      // Same #24 safety net as the router: socketPath might point at a
-      // directory the PG manager has since cleaned up. Fall back to TCP
-      // rather than hanging on a missing socket file.
-      const useUnix = pgSocketPath && fs.existsSync(pgSocketPath);
-      if (useUnix) {
-        state.pgSocket = await Bun.connect({ unix: pgSocketPath, socket: pgHandler });
-      } else {
-        if (pgSocketPath && !useUnix) {
-          this.logger.warn?.(
-            { pgSocketPath, dbName },
-            'PG Unix socket path stale — falling back to TCP',
-          );
-        }
-        state.pgSocket = await Bun.connect({
-          hostname: '127.0.0.1',
-          port: this.pgManager.port,
-          socket: pgHandler,
-        });
-      }
-
-      this.emit('connection', { dbName, socket });
-    } catch (err) {
-      this.logger.error?.({ dbName: state.dbName, err: err?.message || String(err) }, 'Daemon connection error');
-      try { socket.end(); } catch { /* swallow */ }
-      this.emit('connection-error', { error: err, dbName: state.dbName });
-    } finally {
-      state.startupInProgress = false;
-    }
-  }
-
-  /**
-   * Group 4 — wire identity to tenancy.
-   *
-   * On first connect: provision a per-fingerprint database, audit the create.
-   * On reconnect: bump last_connection_at + liveness_pid.
-   * On cross-tenant attempt: deny with SQLSTATE 28P01, OR (when the kill
-   * switch is on) bypass and audit `enforcement_kill_switch_used`.
-   *
-   * Returns either `{databaseName}` (proceed with that name) or
-   * `{deny: true, message, ownedDatabaseName}` (caller writes ErrorResponse
-   * and closes).
-   *
-   * @param {object} state — socket state holding the fingerprint info
-   * @param {string} requestedDb — database name extracted from startup msg
-   */
-  async resolveTenantDatabase(state, requestedDb) {
-    const fp = state.fingerprint;
-    // No fingerprint (FFI unavailable, accept hook failed) or no admin
-    // client (init failed) → behave as v1: route the requested name through
-    // unchanged. Tenancy enforcement is best-effort, never load-bearing for
-    // basic connectivity.
-    if (!fp || !this._adminClient) {
-      return { databaseName: requestedDb };
-    }
-
-    const { fingerprint, name, uid, pid, packageRealpath } = fp;
-
-    let row = null;
-    try {
-      row = await findRowByFingerprint(this._adminClient, fingerprint);
-    } catch (err) {
-      this.logger.warn?.(
-        { err: err?.message || String(err), fingerprint },
-        'pgserve_meta lookup failed — falling back to requested DB',
-      );
-      return { databaseName: requestedDb };
-    }
-
-    if (!row) {
-      const newName = resolveTenantDatabaseName({ name, fingerprint });
-      try {
-        await this.pgManager.createDatabase(newName);
-        await recordDbCreated(this._adminClient, {
-          databaseName: newName,
-          fingerprint,
-          peerUid: typeof uid === 'number' ? uid : -1,
-          packageRealpath: packageRealpath || null,
-          livenessPid: typeof pid === 'number' && pid > 0 ? pid : null,
-          persist: false,
-        });
-        audit(AUDIT_EVENTS.DB_CREATED, {
-          database: newName,
-          fingerprint,
-          peer_uid: uid,
-          peer_pid: pid,
-          package_realpath: packageRealpath || null,
-          name,
-        });
-      } catch (err) {
-        this.logger.error?.(
-          { err: err?.message || String(err), fingerprint, dbName: newName },
-          'Failed to provision per-fingerprint database',
-        );
-        // Bubble — the catch block in processStartupMessage closes the socket.
-        throw err;
-      }
-      row = { databaseName: newName, fingerprint, peerUid: uid, allowedTokens: [] };
-    } else {
-      try {
-        await touchLastConnection(this._adminClient, {
-          databaseName: row.databaseName,
-          livenessPid: typeof pid === 'number' && pid > 0 ? pid : null,
-        });
-      } catch (err) {
-        this.logger.warn?.(
-          { err: err?.message || String(err), database: row.databaseName },
-          'touchLastConnection failed (non-fatal)',
-        );
-      }
-    }
-
-    // Enforcement: peer asked for an explicit database that isn't theirs.
-    // libpq's default `database = user` is treated the same as `postgres` —
-    // both are "I don't care, give me whatever you have for me", so we
-    // silently route them into the fingerprint's DB.
-    const requested = (typeof requestedDb === 'string' && requestedDb.length > 0)
-      ? requestedDb
-      : 'postgres';
-    const isImplicit = requested === 'postgres' || requested === row.databaseName;
-    if (!isImplicit) {
-      if (this.enforcementDisabled) {
-        audit(AUDIT_EVENTS.ENFORCEMENT_KILL_SWITCH_USED, {
-          fingerprint,
-          peer_uid: uid,
-          peer_pid: pid,
-          requested_database: requested,
-          owned_database: row.databaseName,
-        });
-        // Kill switch: route the peer wherever they asked, just like v1.
-        try { await this.pgManager.createDatabase(requested); } catch { /* swallow */ }
-        return { databaseName: requested };
-      }
-      audit(AUDIT_EVENTS.CONNECTION_DENIED_FINGERPRINT_MISMATCH, {
-        fingerprint,
-        peer_uid: uid,
-        peer_pid: pid,
-        requested_database: requested,
-        owned_database: row.databaseName,
-      });
-      return {
-        deny: true,
-        ownedDatabaseName: row.databaseName,
-        message:
-          `database fingerprint mismatch: peer ${fingerprint} owns ` +
-          `${row.databaseName}, requested ${requested}`,
-      };
-    }
-
-    return { databaseName: row.databaseName };
-  }
-
-  handleSocketClose(socket) {
-    const state = this.socketState.get(socket);
-    if (state) {
-      state.pendingToPg = null;
-      state.pendingToClient = null;
-      if (state.pgSocket) {
-        try { state.pgSocket.end(); } catch { /* swallow */ }
-      }
-    }
-    this.connections.delete(socket);
-    this.socketState.delete(socket);
-  }
-
-  handleSocketError(socket, error) {
-    const state = this.socketState.get(socket);
-    if (error?.code !== 'ECONNRESET') {
-      this.logger.error?.({ err: error?.message || String(error), dbName: state?.dbName }, 'Control socket error');
-    }
-    if (state) {
-      state.pendingToPg = null;
-      state.pendingToClient = null;
-      if (state.pgSocket) {
-        try { state.pgSocket.end(); } catch { /* swallow */ }
-      }
-    }
-    this.connections.delete(socket);
-    this.socketState.delete(socket);
-  }
-
   getStats() {
     return {
       controlSocketPath: this.controlSocketPath,
@@ -914,12 +542,11 @@ export class PgserveDaemon extends EventEmitter {
   }
 }
 
-function flushPending(target, pending) {
-  const written = target.write(pending);
-  if (written === pending.byteLength) return null;
-  if (written === 0) return pending;
-  return pending.subarray(written);
-}
+// Mix the accept-path handlers (Unix + TCP) into the prototype. Done at
+// module load so `new PgserveDaemon()` always has them — same observable
+// surface as the pre-split file.
+attachControlHandlers(PgserveDaemon);
+attachTcpHandlers(PgserveDaemon);
 
 /**
  * Normalise the `--listen` form. Accepts:
@@ -959,268 +586,6 @@ function parseSingleListen(spec) {
   }
   return { host: host || '0.0.0.0', port };
 }
-
-/**
- * Add TCP-bind + auth flow methods to PgserveDaemon. These hang off the
- * prototype rather than re-opening the class so the diff reads as one
- * cohesive Group-6 surface.
- */
-PgserveDaemon.prototype.bindTcpListener = async function bindTcpListener({ host, port }) {
-  const daemon = this;
-  return Bun.listen({
-    hostname: host,
-    port,
-    socket: {
-      data(socket, data) { daemon.handleTcpData(socket, data); },
-      open(socket) { daemon.handleTcpOpen(socket); },
-      close(socket) { daemon.handleTcpClose(socket); },
-      error(socket, error) { daemon.handleTcpError(socket, error); },
-      drain(socket) {
-        const state = daemon.socketState.get(socket);
-        if (!state) return;
-        if (state.pendingToClient) {
-          state.pendingToClient = flushPending(socket, state.pendingToClient);
-        }
-        if (!state.pendingToClient && state.pgSocket) {
-          state.pgSocket.resume();
-        }
-      },
-    },
-  });
-};
-
-PgserveDaemon.prototype.handleTcpOpen = function handleTcpOpen(socket) {
-  // TCP peers cannot use SO_PEERCRED; identity is established via the
-  // application_name token in the startup message.
-  this.socketState.set(socket, {
-    transport: 'tcp',
-    buffer: null,
-    pgSocket: null,
-    dbName: null,
-    handshakeComplete: false,
-    startupInProgress: false,
-    pendingToPg: null,
-    pendingToClient: null,
-    fingerprint: null,
-    tokenId: null,
-  });
-  this.connections.add(socket);
-};
-
-PgserveDaemon.prototype.handleTcpData = function handleTcpData(socket, data) {
-  const state = this.socketState.get(socket);
-  if (!state) return;
-
-  if (state.handshakeComplete && state.pgSocket) {
-    if (state.pendingToPg) {
-      state.pendingToPg = Buffer.concat([state.pendingToPg, Buffer.from(data)]);
-      socket.pause();
-      return;
-    }
-    const written = state.pgSocket.write(data);
-    if (written < data.byteLength) {
-      state.pendingToPg = written === 0 ? Buffer.from(data) : Buffer.from(data.subarray(written));
-      socket.pause();
-    }
-    return;
-  }
-
-  const incomingSize = state.buffer ? state.buffer.length + data.byteLength : data.byteLength;
-  if (incomingSize > MAX_STARTUP_BUFFER_SIZE) {
-    this.logger.warn?.(
-      { incomingSize, limit: MAX_STARTUP_BUFFER_SIZE },
-      'TCP pre-handshake buffer exceeded limit — closing connection',
-    );
-    socket.end();
-    return;
-  }
-  state.buffer = state.buffer ? Buffer.concat([state.buffer, Buffer.from(data)]) : Buffer.from(data);
-
-  this.processTcpStartupMessage(socket, state).catch((err) => {
-    this.logger.error?.({ err: err.message }, 'TCP processStartupMessage failed');
-    try { socket.end(); } catch { /* swallow */ }
-  });
-};
-
-PgserveDaemon.prototype.processTcpStartupMessage = async function processTcpStartupMessage(socket, state) {
-  if (state.startupInProgress) return;
-  const buffer = state.buffer;
-  if (!buffer || buffer.length < 8) return;
-  const messageLength = buffer.readUInt32BE(0);
-  if (buffer.length < messageLength) return;
-  const code = buffer.readUInt32BE(4);
-
-  if (code === SSL_REQUEST_CODE || code === GSSAPI_REQUEST_CODE) {
-    socket.write(Buffer.from('N'));
-    state.buffer = buffer.length > messageLength ? buffer.subarray(messageLength) : null;
-    return;
-  }
-  if (code === CANCEL_REQUEST_CODE) {
-    socket.end();
-    return;
-  }
-  if (code !== PROTOCOL_VERSION_3) {
-    this.logger.warn?.({ code }, 'TCP unsupported protocol version');
-    socket.end();
-    return;
-  }
-
-  const startupMessage = buffer.subarray(0, messageLength);
-  const applicationName = extractApplicationName(startupMessage);
-  const auth = parseTcpAuth(applicationName);
-  state.startupInProgress = true;
-
-  // Validate before opening any PG socket. The denied path emits exactly
-  // one audit event then closes — the peer gets no oracle distinguishing
-  // "unknown fingerprint" from "bad token".
-  let validated = null;
-  try {
-    if (auth && this._adminClient) {
-      const tokenHash = hashToken(auth.token);
-      validated = await verifyToken(this._adminClient, {
-        fingerprint: auth.fingerprint,
-        tokenHash,
-      });
-    }
-  } catch (err) {
-    this.logger.warn?.({ err: err.message }, 'verifyToken failed');
-    validated = null;
-  }
-
-  if (!validated) {
-    audit(AUDIT_EVENTS.TCP_TOKEN_DENIED, {
-      fingerprint: auth?.fingerprint || null,
-      remote_address: socket.remoteAddress || null,
-      reason: !auth ? 'missing_or_malformed_application_name' : 'token_unknown',
-    });
-    try { socket.end(); } catch { /* swallow */ }
-    state.startupInProgress = false;
-    return;
-  }
-
-  state.fingerprint = auth.fingerprint;
-  state.tokenId = validated.tokenId;
-  state.dbName = validated.databaseName;
-
-  audit(AUDIT_EVENTS.TCP_TOKEN_USED, {
-    fingerprint: auth.fingerprint,
-    token_id: validated.tokenId,
-    database: validated.databaseName,
-    remote_address: socket.remoteAddress || null,
-  });
-
-  // Force the peer into its fingerprint's database — even if the libpq
-  // client asked for something else. Drop application_name on the way
-  // through: the auth blob easily exceeds Postgres' 63-char NAMEDATALEN
-  // and would otherwise trigger a truncation NOTICE on every connect.
-  let outgoingStartup;
-  try {
-    outgoingStartup = rewriteDatabaseName(startupMessage, validated.databaseName, {
-      dropParams: ['application_name'],
-    });
-  } catch (err) {
-    this.logger.error?.({ err: err.message }, 'rewriteDatabaseName failed for TCP peer');
-    try { socket.end(); } catch { /* swallow */ }
-    state.startupInProgress = false;
-    return;
-  }
-
-  try {
-    if (this.autoProvision) {
-      await this.pgManager.createDatabase(validated.databaseName);
-    }
-    const pgSocketPath = this.pgManager.getSocketPath();
-    const daemon = this;
-    const pgHandler = {
-      data(_pgSocket, pgData) {
-        if (state.pendingToClient) {
-          state.pendingToClient = Buffer.concat([state.pendingToClient, Buffer.from(pgData)]);
-          _pgSocket.pause();
-          return;
-        }
-        const written = socket.write(pgData);
-        if (written < pgData.byteLength) {
-          state.pendingToClient = written === 0
-            ? Buffer.from(pgData)
-            : Buffer.from(pgData.subarray(written));
-          _pgSocket.pause();
-        }
-      },
-      open(pgSocket) {
-        pgSocket.write(outgoingStartup);
-        state.handshakeComplete = true;
-      },
-      close() {
-        try { socket.end(); } catch { /* swallow */ }
-      },
-      error(_pgSocket, error) {
-        daemon.logger.error?.(
-          { dbName: validated.databaseName, err: error?.message || String(error) },
-          'TCP-side PG proxy socket error',
-        );
-        try { socket.end(); } catch { /* swallow */ }
-      },
-      drain(_pgSocket) {
-        if (state.pendingToPg) {
-          state.pendingToPg = flushPending(_pgSocket, state.pendingToPg);
-        }
-        if (!state.pendingToPg) {
-          socket.resume();
-        }
-      },
-    };
-
-    const useUnix = pgSocketPath && fs.existsSync(pgSocketPath);
-    if (useUnix) {
-      state.pgSocket = await Bun.connect({ unix: pgSocketPath, socket: pgHandler });
-    } else {
-      state.pgSocket = await Bun.connect({
-        hostname: '127.0.0.1',
-        port: this.pgManager.port,
-        socket: pgHandler,
-      });
-    }
-    this.emit('tcp-connection', { dbName: validated.databaseName, fingerprint: auth.fingerprint });
-  } catch (err) {
-    this.logger.error?.(
-      { dbName: validated.databaseName, err: err?.message || String(err) },
-      'TCP daemon connection error',
-    );
-    try { socket.end(); } catch { /* swallow */ }
-    this.emit('connection-error', { error: err, dbName: validated.databaseName });
-  } finally {
-    state.startupInProgress = false;
-  }
-};
-
-PgserveDaemon.prototype.handleTcpClose = function handleTcpClose(socket) {
-  const state = this.socketState.get(socket);
-  if (state) {
-    state.pendingToPg = null;
-    state.pendingToClient = null;
-    if (state.pgSocket) {
-      try { state.pgSocket.end(); } catch { /* swallow */ }
-    }
-  }
-  this.connections.delete(socket);
-  this.socketState.delete(socket);
-};
-
-PgserveDaemon.prototype.handleTcpError = function handleTcpError(socket, error) {
-  if (error?.code !== 'ECONNRESET') {
-    this.logger.error?.({ err: error?.message || String(error) }, 'TCP socket error');
-  }
-  const state = this.socketState.get(socket);
-  if (state) {
-    state.pendingToPg = null;
-    state.pendingToClient = null;
-    if (state.pgSocket) {
-      try { state.pgSocket.end(); } catch { /* swallow */ }
-    }
-  }
-  this.connections.delete(socket);
-  this.socketState.delete(socket);
-};
 
 /**
  * Convenience entry — used by the CLI subcommand.
